+++ outputs/logstash/client.go
@@ -24,17 +24,15 @@ type lumberjackClient struct {
 	TransportClient
 	windowSize      int
 	maxOkWindowSize int // max window size sending was successful for
+	maxWindowSize   int
 	timeout         time.Duration
 	countTimeoutErr int
 }
 
-// TODO: make max window size configurable
 const (
 	minWindowSize             int = 1
-	maxWindowSize             int = 1024
 	defaultStartMaxWindowSize int = 10
-
-	maxAllowedTimeoutErr int = 3
+	maxAllowedTimeoutErr      int = 3
 )
 
 // errors
@@ -52,11 +50,16 @@ var (
 	codeCompressed    = []byte{codeVersion, 'C'}
 )
 
-func newLumberjackClient(conn TransportClient, timeout time.Duration) *lumberjackClient {
+func newLumberjackClient(
+	conn TransportClient,
+	maxWindowSize int,
+	timeout time.Duration,
+) *lumberjackClient {
 	return &lumberjackClient{
 		TransportClient: conn,
 		windowSize:      defaultStartMaxWindowSize,
 		timeout:         timeout,
+		maxWindowSize:   maxWindowSize,
 	}
 }
 
@@ -65,10 +68,32 @@ func (l *lumberjackClient) PublishEvent(event common.MapStr) error {
 	return err
 }
 
-func (l *lumberjackClient) PublishEvents(events []common.MapStr) (int, error) {
+// PublishEvents sends all events to logstash. On error a slice with all events
+// not published or confirmed to be processed by logstash will be returned.
+func (l *lumberjackClient) PublishEvents(
+	events []common.MapStr,
+) ([]common.MapStr, error) {
+	for len(events) > 0 {
+		n, err := l.publishWindowed(events)
+
+		logp.Debug("logstash", "%s events out of %s events sent to logstash. Continue sending ...", n, len(events))
+		events = events[n:]
+		if err != nil {
+			return events, err
+		}
+	}
+	return nil, nil
+}
+
+// publishWindowed published events with current maximum window size to logstash
+// returning the total number of events sent (due to window size, or acks until
+// failure).
+func (l *lumberjackClient) publishWindowed(events []common.MapStr) (int, error) {
 	if len(events) == 0 {
 		return 0, nil
 	}
+
+	logp.Debug("logstash", "Try to publish %s events to logstash with window size %s", len(events), l.windowSize)
 
 	// prepare message payload
 	if len(events) > l.windowSize {
@@ -104,7 +129,7 @@ func (l *lumberjackClient) PublishEvents(events []common.MapStr) (int, error) {
 		// read until all acks
 		ackSeq, err = l.readACK()
 		if err != nil {
-			return l.onFail(ackSeq, err)
+			return l.onFail(int(ackSeq), err)
 		}
 	}
 
@@ -114,10 +139,10 @@ func (l *lumberjackClient) PublishEvents(events []common.MapStr) (int, error) {
 	if l.maxOkWindowSize < l.windowSize {
 		l.maxOkWindowSize = l.windowSize
 
-		if l.windowSize < maxWindowSize {
+		if l.windowSize < l.maxWindowSize {
 			l.windowSize = l.windowSize + l.windowSize/2
-			if l.windowSize > maxWindowSize {
-				l.windowSize = maxWindowSize
+			if l.windowSize > l.maxWindowSize {
+				l.windowSize = l.maxWindowSize
 			}
 		}
 	} else if l.windowSize < l.maxOkWindowSize {
@@ -130,30 +155,30 @@ func (l *lumberjackClient) PublishEvents(events []common.MapStr) (int, error) {
 	return len(events), nil
 }
 
-func (l *lumberjackClient) onFail(ackSeq uint32, err error) (int, error) {
+func (l *lumberjackClient) onFail(n int, err error) (int, error) {
 	// if timeout error, back off and ignore error
 	nerr, ok := err.(net.Error)
 	if !ok || !nerr.Timeout() {
 		// no timeout error, close connection and return error
 		_ = l.Close()
-		return int(ackSeq), err
+		return n, err
 	}
 
 	// if we've seen 3 consecutive timeout errors, close connection
 	l.countTimeoutErr++
 	if l.countTimeoutErr == maxAllowedTimeoutErr {
 		_ = l.Close()
-		return int(ackSeq), err
+		return n, err
 	}
 
 	// timeout error. reduce window size and return 0 published events. Send
 	// mode might try to publish again with reduce window size or ask another
-	// client to send events (round robin load balancer)
+	// client to send events
 	l.windowSize = l.windowSize / 2
 	if l.windowSize < minWindowSize {
 		l.windowSize = minWindowSize
 	}
-	return int(ackSeq), nil
+	return n, nil
 }
 
 func (l *lumberjackClient) compressEvents(