+++ outputs/logstash/client_test.go
@@ -82,7 +82,8 @@ func newClientTestDriver(client mode.ProtocolClient) *testClientDriver {
 				close(driver.ch)
 				return
 			case driverCmdPublish:
-				n, err := driver.client.PublishEvents(cmd.events)
+				events, err := driver.client.PublishEvents(cmd.events)
+				n := len(cmd.events) - len(events)
 				driver.returns = append(driver.returns, testClientReturn{n, err})
 			}
 		}
@@ -295,9 +296,12 @@ func sendAck(transp *mockTransport, seq uint32) {
 	transp.sendBytes(buf.Bytes())
 }
 
+const testMaxWindowSize = 64
+
 func TestSendZero(t *testing.T) {
 	transp := newMockTransport()
-	client := newClientTestDriver(newLumberjackClient(transp, 5*time.Second))
+	client := newClientTestDriver(
+		newLumberjackClient(transp, testMaxWindowSize, 5*time.Second))
 
 	client.Publish(make([]common.MapStr, 0))
 
@@ -311,7 +315,8 @@ func TestSendZero(t *testing.T) {
 
 func TestSimpleEvent(t *testing.T) {
 	transp := newMockTransport()
-	client := newClientTestDriver(newLumberjackClient(transp, 5*time.Second))
+	client := newClientTestDriver(
+		newLumberjackClient(transp, testMaxWindowSize, 5*time.Second))
 
 	event := common.MapStr{"name": "me", "line": 10}
 	client.Publish([]common.MapStr{event})
@@ -343,7 +348,8 @@ func TestSimpleEvent(t *testing.T) {
 
 func TestStructuredEvent(t *testing.T) {
 	transp := newMockTransport()
-	client := newClientTestDriver(newLumberjackClient(transp, 5*time.Second))
+	client := newClientTestDriver(
+		newLumberjackClient(transp, testMaxWindowSize, 5*time.Second))
 	event := common.MapStr{
 		"name": "test",
 		"struct": common.MapStr{