+++ outputs/logstash/logstash.go
@@ -5,8 +5,6 @@ package logstash
 
 import (
 	"crypto/tls"
-	"errors"
-	"fmt"
 	"time"
 
 	"github.com/elastic/libbeat/common"
@@ -44,25 +42,23 @@ type logstash struct {
 const (
 	logstashDefaultPort = 10200
 
-	logstashDefaultTimeout = 30 * time.Second
-	defaultSendRetries     = 3
+	logstashDefaultTimeout   = 30 * time.Second
+	logstasDefaultMaxTimeout = 90 * time.Second
+	defaultSendRetries       = 3
+	defaultMaxWindowSize     = 1024
 )
 
-// ErrNoHostsConfigured indicates missing host or hosts configuration
-var ErrNoHostsConfigured = errors.New("no host configuration found")
-
 var waitRetry = time.Duration(1) * time.Second
+
+// NOTE: maxWaitRetry has no effect on mode, as logstash client currently does not return ErrTempBulkFailure
+var maxWaitRetry = time.Duration(60) * time.Second
 
 func (lj *logstash) init(
 	beat string,
 	config outputs.MothershipConfig,
 	topologyExpire int,
 ) error {
-	useTLS := false
-	if config.TLS != nil {
-		useTLS = !config.TLS.Disabled
-	}
-
+	useTLS := (config.TLS != nil)
 	timeout := logstashDefaultTimeout
 	if config.Timeout != 0 {
 		timeout = time.Duration(config.Timeout) * time.Second
@@ -71,6 +67,11 @@ func (lj *logstash) init(
 	defaultPort := logstashDefaultPort
 	if config.Port != 0 {
 		defaultPort = config.Port
+	}
+
+	maxWindowSize := defaultMaxWindowSize
+	if config.BulkMaxSize != nil {
+		maxWindowSize = *config.BulkMaxSize
 	}
 
 	var clients []mode.ProtocolClient
@@ -82,15 +83,17 @@ func (lj *logstash) init(
 			return err
 		}
 
-		clients, err = makeClients(config, timeout,
-			func(host string) (TransportClient, error) {
-				return newTLSClient(host, defaultPort, tlsConfig)
-			})
+		clients, err = mode.MakeClients(config,
+			makeClientFactory(maxWindowSize, timeout,
+				func(host string) (TransportClient, error) {
+					return newTLSClient(host, defaultPort, tlsConfig)
+				}))
 	} else {
-		clients, err = makeClients(config, timeout,
-			func(host string) (TransportClient, error) {
-				return newTCPClient(host, defaultPort)
-			})
+		clients, err = mode.MakeClients(config,
+			makeClientFactory(maxWindowSize, timeout,
+				func(host string) (TransportClient, error) {
+					return newTCPClient(host, defaultPort)
+				}))
 	}
 	if err != nil {
 		return err
@@ -100,17 +103,22 @@ func (lj *logstash) init(
 	if config.Max_retries != nil {
 		sendRetries = *config.Max_retries
 	}
+	maxAttempts := sendRetries + 1
+	if sendRetries < 0 {
+		maxAttempts = 0
+	}
 
 	var m mode.ConnectionMode
 	if len(clients) == 1 {
 		m, err = mode.NewSingleConnectionMode(clients[0],
-			sendRetries, waitRetry, timeout)
+			maxAttempts, waitRetry, timeout, maxWaitRetry)
 	} else {
 		loadBalance := config.LoadBalance != nil && *config.LoadBalance
 		if loadBalance {
-			m, err = mode.NewLoadBalancerMode(clients, sendRetries, waitRetry, timeout)
+			m, err = mode.NewLoadBalancerMode(clients, maxAttempts,
+				waitRetry, timeout, maxWaitRetry)
 		} else {
-			m, err = mode.NewFailOverConnectionMode(clients, sendRetries, waitRetry, timeout)
+			m, err = mode.NewFailOverConnectionMode(clients, maxAttempts, waitRetry, timeout)
 		}
 	}
 	if err != nil {
@@ -126,34 +134,17 @@ func (lj *logstash) init(
 	return nil
 }
 
-func makeClients(
-	config outputs.MothershipConfig,
+func makeClientFactory(
+	maxWindowSize int,
 	timeout time.Duration,
-	newTransp func(string) (TransportClient, error),
-) ([]mode.ProtocolClient, error) {
-	switch {
-	case len(config.Hosts) > 0:
-		var clients []mode.ProtocolClient
-		for _, host := range config.Hosts {
-			transp, err := newTransp(host)
-			if err != nil {
-				for _, client := range clients {
-					_ = client.Close() // ignore error
-				}
-				return nil, err
-			}
-			client := newLumberjackClient(transp, timeout)
-			clients = append(clients, client)
-		}
-		return clients, nil
-	case config.Host != "":
-		transp, err := newTransp(config.Host)
+	makeTransp func(string) (TransportClient, error),
+) func(string) (mode.ProtocolClient, error) {
+	return func(host string) (mode.ProtocolClient, error) {
+		transp, err := makeTransp(host)
 		if err != nil {
 			return nil, err
 		}
-		return []mode.ProtocolClient{newLumberjackClient(transp, timeout)}, nil
-	default:
-		return nil, ErrNoHostsConfigured
+		return newLumberjackClient(transp, maxWindowSize, timeout), nil
 	}
 }
 
@@ -187,11 +178,8 @@ func (lj *logstash) BulkPublish(
 // decode/rename the "line" field into "message".
 func (lj *logstash) addMeta(event common.MapStr) {
 	// add metadata for indexing
-	ts := time.Time(event["timestamp"].(common.Time)).UTC()
-	index := fmt.Sprintf("%s-%02d.%02d.%02d", lj.index,
-		ts.Year(), ts.Month(), ts.Day())
 	event["@metadata"] = common.MapStr{
-		"index": index,
-		"type":  event["type"].(string),
+		"beat": lj.index,
+		"type": event["type"].(string),
 	}
 }