+++ outputs/mode/mode_test.go
@@ -1,6 +1,7 @@
 package mode
 
 import (
+	"errors"
 	"sync"
 	"testing"
 	"time"
@@ -11,7 +12,7 @@ import (
 )
 
 type mockClient struct {
-	publish   func([]common.MapStr) (int, error)
+	publish   func([]common.MapStr) ([]common.MapStr, error)
 	close     func() error
 	connected bool
 	connect   func(time.Duration) error
@@ -29,7 +30,7 @@ func (c *mockClient) IsConnected() bool {
 	return c.connected
 }
 
-func (c *mockClient) PublishEvents(events []common.MapStr) (int, error) {
+func (c *mockClient) PublishEvents(events []common.MapStr) ([]common.MapStr, error) {
 	return c.publish(events)
 }
 
@@ -62,14 +63,14 @@ func alwaysFailConnect(err error) func(time.Duration) error {
 
 func collectPublish(
 	collected *[][]common.MapStr,
-) func(events []common.MapStr) (int, error) {
+) func(events []common.MapStr) ([]common.MapStr, error) {
 	mutex := sync.Mutex{}
-	return func(events []common.MapStr) (int, error) {
+	return func(events []common.MapStr) ([]common.MapStr, error) {
 		mutex.Lock()
 		defer mutex.Unlock()
 
 		*collected = append(*collected, events)
-		return len(events), nil
+		return nil, nil
 	}
 }
 
@@ -81,32 +82,40 @@ func (e errNetTimeout) Temporary() bool { return false }
 
 func publishTimeoutEvery(
 	n int,
-	pub func(events []common.MapStr) (int, error),
-) func(events []common.MapStr) (int, error) {
+	pub func(events []common.MapStr) ([]common.MapStr, error),
+) func(events []common.MapStr) ([]common.MapStr, error) {
 	count := 0
-	return func(events []common.MapStr) (int, error) {
+	return func(events []common.MapStr) ([]common.MapStr, error) {
 		if count < n {
 			count++
 			return pub(events)
 		}
 		count = 0
-		return 0, errNetTimeout{}
+		return events, errNetTimeout{}
+	}
+}
+
+func publishFailWith(
+	n int,
+	err error,
+	pub func([]common.MapStr) ([]common.MapStr, error),
+) func([]common.MapStr) ([]common.MapStr, error) {
+	count := 0
+	return func(events []common.MapStr) ([]common.MapStr, error) {
+		if count < n {
+			count++
+			return events, err
+		}
+		count = 0
+		return pub(events)
 	}
 }
 
 func publishFailStart(
 	n int,
-	pub func(events []common.MapStr) (int, error),
-) func(events []common.MapStr) (int, error) {
-	count := 0
-	return func(events []common.MapStr) (int, error) {
-		if count < n {
-			count++
-			return 0, errNetTimeout{}
-		}
-		count = 0
-		return pub(events)
-	}
+	pub func(events []common.MapStr) ([]common.MapStr, error),
+) func(events []common.MapStr) ([]common.MapStr, error) {
+	return publishFailWith(n, errNetTimeout{}, pub)
 }
 
 func closeOK() error {
@@ -207,4 +216,78 @@ func repeat(n int, evt []eventInfo) []eventInfo {
 
 func signals(s ...bool) []bool {
 	return s
+}
+
+func TestMakeEmptyClientFail(t *testing.T) {
+	config := outputs.MothershipConfig{}
+	clients, err := MakeClients(config, dummyMockClientFactory)
+	assert.Equal(t, ErrNoHostsConfigured, err)
+	assert.Equal(t, 0, len(clients))
+}
+
+func TestMakeSingleClient(t *testing.T) {
+	config := outputs.MothershipConfig{
+		Hosts: []string{"single"},
+	}
+
+	clients, err := MakeClients(config, dummyMockClientFactory)
+	assert.Nil(t, err)
+	assert.Equal(t, 1, len(clients))
+}
+
+func TestMakeSingleClientWorkers(t *testing.T) {
+	config := outputs.MothershipConfig{
+		Hosts:  []string{"single"},
+		Worker: 3,
+	}
+
+	clients, err := MakeClients(config, dummyMockClientFactory)
+	assert.Nil(t, err)
+	assert.Equal(t, 3, len(clients))
+}
+
+func TestMakeTwoClient(t *testing.T) {
+	config := outputs.MothershipConfig{
+		Hosts: []string{"client1", "client2"},
+	}
+
+	clients, err := MakeClients(config, dummyMockClientFactory)
+	assert.Nil(t, err)
+	assert.Equal(t, 2, len(clients))
+}
+
+func TestMakeTwoClientWorkers(t *testing.T) {
+	config := outputs.MothershipConfig{
+		Hosts:  []string{"client1", "client2"},
+		Worker: 3,
+	}
+
+	clients, err := MakeClients(config, dummyMockClientFactory)
+	assert.Nil(t, err)
+	assert.Equal(t, 6, len(clients))
+}
+
+func TestMakeTwoClientFail(t *testing.T) {
+	config := outputs.MothershipConfig{
+		Hosts:  []string{"client1", "client2"},
+		Worker: 3,
+	}
+
+	testError := errors.New("test")
+
+	i := 1
+	_, err := MakeClients(config, func(host string) (ProtocolClient, error) {
+		if i%3 == 0 {
+			return nil, testError
+		}
+		i++
+		return dummyMockClientFactory(host)
+	})
+	assert.Equal(t, testError, err)
+}
+
+func dummyMockClientFactory(host string) (ProtocolClient, error) {
+	return &mockClient{
+		close: func() error { return nil },
+	}, nil
 }