+++ outputs/mode/single.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/elastic/libbeat/common"
+	"github.com/elastic/libbeat/logp"
 	"github.com/elastic/libbeat/outputs"
 )
 
@@ -15,8 +16,9 @@ type SingleConnectionMode struct {
 
 	closed bool // mode closed flag to break publisher loop
 
-	timeout   time.Duration // connection timeout
-	waitRetry time.Duration // wait time until reconnect
+	timeout      time.Duration // connection timeout
+	waitRetry    time.Duration // wait time until reconnect
+	maxWaitRetry time.Duration // Maximum send/retry timeout in backoff case.
 
 	// maximum number of configured send attempts. If set to 0, publisher will
 	// block until event has been successfully published.
@@ -28,11 +30,15 @@ type SingleConnectionMode struct {
 func NewSingleConnectionMode(
 	client ProtocolClient,
 	maxAttempts int,
-	waitRetry, timeout time.Duration,
+	waitRetry, timeout, maxWaitRetry time.Duration,
 ) (*SingleConnectionMode, error) {
 	s := &SingleConnectionMode{
-		timeout:     timeout,
-		conn:        client,
+		conn: client,
+
+		timeout:      timeout,
+		waitRetry:    waitRetry,
+		maxWaitRetry: maxWaitRetry,
+
 		maxAttempts: maxAttempts,
 	}
 
@@ -59,36 +65,22 @@ func (s *SingleConnectionMode) PublishEvents(
 	signaler outputs.Signaler,
 	events []common.MapStr,
 ) error {
-	published := 0
-	fails := 0
-	for !s.closed && (s.maxAttempts == 0 || fails < s.maxAttempts) {
-		if err := s.connect(); err != nil {
-			fails++
-			time.Sleep(s.waitRetry)
-			continue
-		}
+	return s.publish(signaler, func() (bool, bool) {
+		for len(events) > 0 {
+			var err error
 
-		for published < len(events) {
-			n, err := s.conn.PublishEvents(events[published:])
+			total := len(events)
+			events, err = s.conn.PublishEvents(events)
 			if err != nil {
-				break
+				logp.Info("Error publishing events (retrying): %s", err)
+
+				madeProgress := len(events) < total
+				return false, madeProgress
 			}
-
-			fails = 0
-			published += n
 		}
 
-		if published == len(events) {
-			outputs.SignalCompleted(signaler)
-			return nil
-		}
-
-		time.Sleep(s.waitRetry)
-		fails++
-	}
-
-	outputs.SignalFailed(signaler)
-	return nil
+		return true, false
+	})
 }
 
 // PublishEvent forwards a single event. On failure PublishEvent tries to reconnect.
@@ -96,22 +88,69 @@ func (s *SingleConnectionMode) PublishEvent(
 	signaler outputs.Signaler,
 	event common.MapStr,
 ) error {
-	fails := 0
-	for !s.closed && (s.maxAttempts == 0 || fails < s.maxAttempts) {
-		if err := s.connect(); err != nil {
-			fails++
-			time.Sleep(s.waitRetry)
-			continue
-		}
+	return s.publish(signaler, func() (bool, bool) {
 		if err := s.conn.PublishEvent(event); err != nil {
-			fails++
-			continue
+			logp.Info("Error publishing event (retrying): %s", err)
+			return false, false
+		}
+		return true, false
+	})
+}
+
+// publish is used to publish events using the configured protocol client.
+// It provides general error handling and back off support used on failed
+// send attempts. To be used by PublishEvent and PublishEvents.
+// The send callback will try to progress sending traffic and returns kind of
+// progress made in ok or resetFail. If ok is set to true, send finished
+// processing events. If ok is false but resetFail is set, send was partially
+// successful. If send was partially successful, the fail counter is reset thus up
+// to maxAttempts send attempts without any progress might be executed.
+func (s *SingleConnectionMode) publish(
+	signaler outputs.Signaler,
+	send func() (ok bool, resetFail bool),
+) error {
+	fails := 0
+	var backoffCount uint
+	var err error
+
+	for !s.closed && (s.maxAttempts == 0 || fails < s.maxAttempts) {
+		ok := false
+		resetFail := false
+
+		if err := s.connect(); err != nil {
+			logp.Info("Connecting error publishing events (retrying): %s", err)
+			goto sendFail
+		}
+
+		ok, resetFail = send()
+		if !ok {
+			goto sendFail
 		}
 
 		outputs.SignalCompleted(signaler)
 		return nil
+
+	sendFail:
+		fails++
+		if resetFail {
+			fails = 0
+		}
+		if s.maxAttempts > 0 && fails == s.maxAttempts {
+			// max number of attempts reached
+			break
+		}
+
+		logp.Info("send fail")
+		backoff := time.Duration(int64(s.waitRetry) * (1 << backoffCount))
+		if backoff > s.maxWaitRetry {
+			backoff = s.maxWaitRetry
+		} else {
+			backoffCount++
+		}
+		logp.Info("backoff retry: %v", backoff)
+		time.Sleep(backoff)
 	}
 
-	outputs.SignalFailed(signaler)
+	outputs.SignalFailed(signaler, err)
 	return nil
 }