+++ publisher/publish.go
@@ -13,6 +13,7 @@ import (
 	"github.com/nranchev/go-libGeoIP"
 
 	// load supported output plugins
+	_ "github.com/elastic/libbeat/outputs/console"
 	_ "github.com/elastic/libbeat/outputs/elasticsearch"
 	_ "github.com/elastic/libbeat/outputs/fileout"
 	_ "github.com/elastic/libbeat/outputs/logstash"
@@ -26,8 +27,18 @@ var debug = logp.MakeDebug("publish")
 
 // EventPublisher provides the interface for beats to publish events.
 type eventPublisher interface {
-	PublishEvent(event common.MapStr) bool
-	PublishEvents(events []common.MapStr) bool
+	PublishEvent(ctx *context, event common.MapStr) bool
+	PublishEvents(ctx *context, events []common.MapStr) bool
+}
+
+type context struct {
+	publishOptions
+	signal outputs.Signaler
+}
+
+type publishOptions struct {
+	confirm bool
+	sync    bool
 }
 
 type TransactionalEventPublisher interface {
@@ -35,7 +46,9 @@ type TransactionalEventPublisher interface {
 }
 
 type PublisherType struct {
-	name           string
+	shipperName    string // Shipper name as set in the configuration file
+	hostname       string // Host name as returned by the operation system
+	name           string // The shipperName if configured, the hostname otherwise
 	ipaddrs        []string
 	tags           []string
 	disabled       bool
@@ -153,7 +166,7 @@ func (publisher *PublisherType) PublishTopology(params ...string) error {
 }
 
 func (publisher *PublisherType) Init(
-	beat string,
+	beatName string,
 	configs map[string]outputs.MothershipConfig,
 	shipper ShipperConfig,
 ) error {
@@ -171,7 +184,7 @@ func (publisher *PublisherType) Init(
 	publisher.wsPublisher.Init()
 
 	if !publisher.disabled {
-		plugins, err := outputs.InitOutputs(beat, configs, shipper.Topology_expire)
+		plugins, err := outputs.InitOutputs(beatName, configs, shipper.Topology_expire)
 		if err != nil {
 			return err
 		}
@@ -182,7 +195,7 @@ func (publisher *PublisherType) Init(
 			output := plugin.Output
 			config := plugin.Config
 
-			debug("create output worker: %p, %p", config.Flush_interval, config.Bulk_size)
+			debug("create output worker: %p, %p", config.Flush_interval, config.BulkMaxSize)
 
 			outputers = append(outputers,
 				newOutputWorker(config, output, &publisher.wsOutput, 1000))
@@ -223,16 +236,17 @@ func (publisher *PublisherType) Init(
 		}
 	}
 
-	publisher.name = shipper.Name
-	if len(publisher.name) == 0 {
-		// use the hostname
-		publisher.name, err = os.Hostname()
-		if err != nil {
-			return err
-		}
-
-		logp.Info("No shipper name configured, using hostname '%s'", publisher.name)
+	publisher.shipperName = shipper.Name
+	publisher.hostname, err = os.Hostname()
+	if err != nil {
+		return err
 	}
+	if len(publisher.shipperName) > 0 {
+		publisher.name = publisher.shipperName
+	} else {
+		publisher.name = publisher.hostname
+	}
+	logp.Info("Publisher name: %s", publisher.name)
 
 	publisher.tags = shipper.Tags